Fusion SwarmInt (solver non fonctionnel)
