--- Sudoku.SwarmInt/SudokuCombinatorialEvolutionSolver/SudokuCombinatorialEvolutionSolver.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.808.5
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SudokuCombinatorialEvolutionSolver", "SudokuCombinatorialEvolutionSolver.csproj", "{2B7B4A69-E672-4CC9-895B-67FD6469A8AC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2B7B4A69-E672-4CC9-895B-67FD6469A8AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2B7B4A69-E672-4CC9-895B-67FD6469A8AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2B7B4A69-E672-4CC9-895B-67FD6469A8AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2B7B4A69-E672-4CC9-895B-67FD6469A8AC}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0DB428E6-5355-41D3-B223-E8C6F4E3B827}
-	EndGlobalSection
-EndGlobal
