--- sudoku-combinatorial-evolution-solver-master/README.md
@@ -1,4 +1,0 @@
-# Solving Sudoku Using Combinatorial Evolution Algorithm
-
-## License
-MIT
