--- sudoku-combinatorial-evolution-solver-master/src/SudokuCombinatorialEvolutionSolver.sln
@@ -1,16 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SudokuCombinatorialEvolutionSolver", "SudokuCombinatorialEvolutionSolver\SudokuCombinatorialEvolutionSolver.csproj", "{E085F87B-8F17-4033-8157-F43C22DA50E5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E085F87B-8F17-4033-8157-F43C22DA50E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E085F87B-8F17-4033-8157-F43C22DA50E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E085F87B-8F17-4033-8157-F43C22DA50E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E085F87B-8F17-4033-8157-F43C22DA50E5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
