--- sudoku-combinatorial-evolution-solver-master/src/SudokuCombinatorialEvolutionSolver/OrganismType.cs
@@ -1,8 +1,0 @@
-namespace SudokuCombinatorialEvolutionSolver
-{
-  public enum OrganismType
-  {
-    Worker,
-    Explorer
-  }
-}
