--- sudoku-combinatorial-evolution-solver-master/src/SudokuCombinatorialEvolutionSolver/SudokuCombinatorialEvolutionSolver.csproj
@@ -1,6 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
-</Project>
